feat: use MediaStore API to delete source file

+++ app/src/main/java/remix/myplayer/bean/mp3/Song.kt
@@ -69,7 +69,7 @@ data class Song(
     }
 
   val contentUri: Uri
-    get() = ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, id.toLong())
+    get() = ContentUris.withAppendedId(MediaStore.Audio.Media.EXTERNAL_CONTENT_URI, id)
 
   val artUri: Uri
     get() = ContentUris.withAppendedId(Uri.parse("content://media/external/audio/albumart/"), albumId)
